src/main/java/br/com/natanmaia/converter/DozerConverter.java
package br.com.natanmaia.converter;

import java.util.ArrayList;
import java.util.List;

import com.github.dozermapper.core.DozerBeanMapperBuilder;
import com.github.dozermapper.core.Mapper;

public class DozerConverter {
	
	private static Mapper mapper = DozerBeanMapperBuilder.buildDefault();
	
	public static <O, D> D parseObject(O origin, Class<D> destination) {
		return mapper.map(origin, destination);
	}
	
	public static <O, D> List<D> parseListObject(List<O> origin, Class<D> destination) {
		List<D> destinationList  = new ArrayList<D>();
		for (Object o : origin) {
			destinationList.add(mapper.map(o, destination));
		}
		return destinationList;
	}

}
